created middleware checking auth and added in handlers posts. added check user is author of post

+++ internal/posts/posts.go
@@ -2,30 +2,34 @@ package posts
 
 import (
 	"errors"
+	"log"
 	"net/http"
 	"strconv"
 
 	"github.com/MLbeL/blog_with_golang/db"
 	"github.com/MLbeL/blog_with_golang/db/models"
+	"github.com/MLbeL/blog_with_golang/pkg/middleware"
 	"github.com/MLbeL/blog_with_golang/pkg/request"
 	"github.com/MLbeL/blog_with_golang/pkg/response"
 )
 
 type PostsHandler struct {
 	PostRepo *db.PostRepo
+	UserRepo *db.UserRepo
 }
 
 type PostsHandlerDeps struct {
 	PostRepo *db.PostRepo
+	UserRepo *db.UserRepo
 }
 
-func NewHandlerPosts(router *http.ServeMux, deps *PostsHandlerDeps) {
-	handler := PostsHandler{PostRepo: deps.PostRepo}
+func NewHandlerPosts(router *http.ServeMux, deps *PostsHandlerDeps, secret string) {
+	handler := PostsHandler{PostRepo: deps.PostRepo, UserRepo: deps.UserRepo}
 	router.HandleFunc("GET /posts", handler.GET_Posts())
-	router.HandleFunc("POST /posts", handler.POST_Posts())
+	router.Handle("POST /posts", middleware.CheckAuth(handler.POST_Posts(), secret))
 	router.HandleFunc("GET /posts/{id}", handler.GETOnePost())
-	router.HandleFunc("PATCH /posts/{id}", handler.UpdatePost())
-	router.HandleFunc("DELETE /posts/{id}", handler.DeletePost())
+	router.Handle("PATCH /posts/{id}", middleware.CheckAuth(handler.UpdatePost(), secret))
+	router.Handle("DELETE /posts/{id}", middleware.CheckAuth(handler.DeletePost(), secret))
 }
 
 func (ph PostsHandler) GET_Posts() http.HandlerFunc {
@@ -46,8 +50,21 @@ func (ph PostsHandler) POST_Posts() http.HandlerFunc {
 		if err != nil {
 			return
 		}
+		email := r.Context().Value(middleware.CtxEmailKey).(string)
+		log.Println(email)
+		_, userID, err := ph.UserRepo.GetHashPasswordByEmailAndUserID(email)
+		log.Println(userID)
+		if err != nil {
+			if errors.Is(err, db.ErrUserIDNotFound) {
+				response.Json("Unauthorized", w, 401)
+				return
+			}
+			response.Json("Internal server error", w, 500)
+			return
+		}
+
 		model_post := models.Post{
-			AuthorID: 1, // TODO: replace with real user id from auth
+			AuthorID: userID,
 			Title:    body.Title,
 			Text:     body.Text,
 		}
@@ -91,6 +108,29 @@ func (ph PostsHandler) UpdatePost() http.HandlerFunc {
 			response.Json("internal server error", w, 500)
 			return
 		}
+
+		// Check if user is author of post
+		email := r.Context().Value(middleware.CtxEmailKey).(string)
+		_, userID, err := ph.UserRepo.GetHashPasswordByEmailAndUserID(email)
+		if err != nil {
+			if errors.Is(err, db.ErrUserIDNotFound) {
+				response.Json("Unauthorized", w, 401)
+				return
+			}
+			response.Json("Internal server error", w, 500)
+			return
+		}
+		post, err := ph.PostRepo.FindPostByID(uint(id))
+		if err != nil {
+			response.Json("internal server error", w, 500)
+			return
+		}
+		if userID != post.AuthorID {
+			response.Json("No permission", w, 403)
+			return
+		}
+		// end of check user is author
+
 		body, err := request.Resp[UpdatePost](&w, r)
 		if err != nil {
 			return
@@ -124,6 +164,33 @@ func (ph PostsHandler) DeletePost() http.HandlerFunc {
 			response.Json("internal server error", w, 500)
 			return
 		}
+
+		// Check if user is author of post
+		email := r.Context().Value(middleware.CtxEmailKey).(string)
+		_, userID, err := ph.UserRepo.GetHashPasswordByEmailAndUserID(email)
+		if err != nil {
+			if errors.Is(err, db.ErrUserIDNotFound) {
+				response.Json("Unauthorized", w, 401)
+				return
+			}
+			response.Json("Internal server error", w, 500)
+			return
+		}
+		post, err := ph.PostRepo.FindPostByID(uint(id))
+		if err != nil {
+			if errors.Is(err, db.ErrPostNotFound) {
+				response.Json("Post not found", w, 404)
+				return
+			}
+			response.Json("Internal server error", w, 500)
+			return
+		}
+		if userID != post.AuthorID {
+			response.Json("No permission", w, 403)
+			return
+		}
+		// end of check user is author
+
 		err = ph.PostRepo.DeletePost(uint(id))
 		if err != nil {
 			if errors.Is(err, db.ErrPostNotFound) {
